[test_fgnhg] Add test to validate FG ECMP packet flow when neighbors are not resolved (#4225)
* Test case enhancement for newly added capability in Fine Grained ECMP to handle packet flow upon no ECMP next-hops being neighbor resolved.

+++ ansible/roles/test/files/ptftests/fg_ecmp_test.py
@@ -33,6 +33,64 @@ IPV4_SRC_IP_RANGE = ['8.0.0.0', '8.255.255.255']
 IPV6_SRC_IP_RANGE = ['20D0:A800:0:00::', '20D0:FFFF:0:00::FFFF']
 
 PERSIST_MAP = '/tmp/fg_ecmp_persist_map.json'
+MAX_ONE_PERCENT_LOSS = 0.01
+
+def verify_packet_warm(test, pkt, port, device_number=0, timeout=None, n_timeout=None):
+    # This packet verification function accounts for possible flood during warm boot
+    # We ensure that packets are received on the expected port, and return a special
+    # return value of -1 to denote that a flood had occured. The caller can use the 
+    # special return value to identify how many packets were flooded. 
+
+    if timeout is None:
+        timeout = ptf.ptfutils.default_timeout
+    if n_timeout is None:
+        n_timeout = ptf.ptfutils.default_negative_timeout
+    logging.debug("Checking for pkt on device %d, port %r", device_number, port)
+    result = dp_poll(test, device_number=device_number, timeout=timeout, exp_pkt=pkt)
+    verify_no_other_packets(test, device_number=device_number, timeout=n_timeout)
+
+    if isinstance(result, test.dataplane.PollSuccess):
+        if result.port != port:
+            # Flood case, check if packet rcvd on expected port as well
+            verify_packet(test, pkt, port)
+            return (-1, None)
+        else:
+            return (port, result.packet)
+
+    assert(isinstance(result, test.dataplane.PollFailure))
+    test.fail("Did not receive expected packet on any of ports %r for device %d.\n%s"
+                % (ports, device_number, result.format()))
+    return (0, None)
+
+def verify_packet_any_port_lossy(test, pkt, ports=[], device_number=0, timeout=None, n_timeout=None):
+    # This packet verification function accounts for possible loss of packet due to route table change
+    # We ensure that packets are received on the expected ports, and return a special
+    # return value of -1 to denote that a packet loss occured. The caller can use the 
+    # special return value to identify how many packets were lost and check if loss is within acceptable range
+
+    if timeout is None:
+        timeout = ptf.ptfutils.default_timeout
+    if n_timeout is None:
+        n_timeout = ptf.ptfutils.default_negative_timeout
+    logging.debug("Checking for pkt on device %d, port %r", device_number, ports)
+    result = dp_poll(test, device_number=device_number, timeout=timeout, exp_pkt=pkt)
+    verify_no_other_packets(test, device_number=device_number, timeout=n_timeout)
+
+    if isinstance(result, test.dataplane.PollSuccess):
+        if result.port in ports:
+            return (ports.index(result.port), result.packet)
+        else:
+            test.fail(
+                "Received expected packet on port %r for device %d, but "
+                "it should have arrived on one of these ports: %r.\n%s"
+                % (result.port, device_number, ports, result.format())
+            )
+            return (0, None)
+
+    if isinstance(result, test.dataplane.PollFailure):
+        return (-1, None)
+
+    return (0, None)
 
 class FgEcmpTest(BaseTest):
 
@@ -49,7 +107,7 @@ class FgEcmpTest(BaseTest):
 
 
     def trigger_mac_learning(self, serv_ports):
-      	for src_port in serv_ports:
+        for src_port in serv_ports:
             pkt = simple_eth_packet(
                 eth_dst=self.router_mac,
                 eth_src=self.dataplane.get_mac(0, src_port),
@@ -151,11 +209,15 @@ class FgEcmpTest(BaseTest):
         tuple_to_port_map ={}
         hit_count_map = {}
 
-        if not os.path.exists(PERSIST_MAP):
+        if not os.path.exists(PERSIST_MAP) and self.test_case == 'create_flows':
             with open(PERSIST_MAP, 'w'): pass
-        else:
+        elif not self.test_case == 'verify_packets_received':
             with open(PERSIST_MAP) as fp:
-                tuple_to_port_map = json.load(fp)
+                try:
+                    tuple_to_port_map = json.load(fp)
+                except ValueError:
+                    print 'Decoding JSON failed for persist map'
+                    assert False
 
         if tuple_to_port_map is None or self.dst_ip not in tuple_to_port_map:
             tuple_to_port_map[self.dst_ip] = {}
@@ -188,6 +250,7 @@ class FgEcmpTest(BaseTest):
                 (port_idx, _) = self.send_rcv_ip_pkt(
                     in_port, src_port, dst_port, src_ip, dst_ip, self.serv_ports, ipv4)
                 assert port_idx == port
+                tuple_to_port_map[self.dst_ip][src_ip] = port_idx
             return
 
         elif self.test_case == 'hash_check_warm_boot':
@@ -199,12 +262,31 @@ class FgEcmpTest(BaseTest):
                 else:
                     in_port = self.net_ports[0]
                 (port_idx, _) = self.send_rcv_ip_pkt_warm(
-                    in_port, src_port, dst_port, src_ip, dst_ip, self.serv_ports, ipv4, port)
+                    in_port, src_port, dst_port, src_ip, dst_ip, port, ipv4)
                 if port_idx == -1:
                     total_flood_pkts = total_flood_pkts + 1
             # Ensure that flooding duration in warm reboot is less than 10% of total packet count
             self.log("Number of flood packets were: " + str(total_flood_pkts))
             assert (total_flood_pkts < (0.1 * len(tuple_to_port_map[self.dst_ip])))
+            return
+
+        elif self.test_case == 'verify_packets_received':
+            self.log("Ensure that all packets were received ...")
+            total_num_pkts_lost = 0
+            for i in range(0, self.num_flows):
+                src_ip = str(base_ip + i)
+                if self.inner_hashing:
+                    in_port = random.choice(self.net_ports)
+                else:
+                    in_port = self.net_ports[0]
+                (port_idx, _) = self.send_rcv_ip_pkt_lossy(
+                    in_port, src_port, dst_port, src_ip, dst_ip, self.serv_ports, ipv4)
+                if port_idx == -1:
+                    total_num_pkts_lost = total_num_pkts_lost + 1
+
+            self.log("Number of lost packets were: " + str(total_num_pkts_lost))
+            # Ensure less than 1% packet loss
+            assert (total_num_pkts_lost < (MAX_ONE_PERCENT_LOSS * self.num_flows))
             return
 
         elif self.test_case == 'bank_check':
@@ -333,36 +415,21 @@ class FgEcmpTest(BaseTest):
         return
 
 
-    def verify_packet_warm(test, pkt, port, device_number=0, timeout=None, n_timeout=None):
-        # This packet verification function accounts for possible flood during warm boot
-        # We ensure that packets are received on the expected port, and return a special
-        # return value of -1 to denote that a flood had occured. The caller can use the 
-        # special return value to identify how many packets were flooded. 
+    def send_rcv_ip_pkt_lossy(self, in_port, sport, dport, src_ip_addr, dst_ip_addr,
+                            exp_port, ipv4=True):
 
-        if timeout is None:
-            timeout = ptf.ptfutils.default_timeout
-        if n_timeout is None:
-            n_timeout = ptf.ptfutils.default_negative_timeout
-        logging.debug("Checking for pkt on device %d, port %r", device_number, port)
-        result = dp_poll(test, device_number=device_number, timeout=timeout, exp_pkt=pkt)
-        verify_no_other_packets(test, device_number=device_number, timeout=n_timeout)
+        if ipv4:
+            (matched_index, received) = self.send_rcv_ipv4_pkt(in_port, sport, dport,
+                    src_ip_addr, dst_ip_addr, exp_port, verify_packet_any_port_lossy)
+        else:
+            (matched_index, received) = self.send_rcv_ipv6_pkt(in_port, sport, dport,
+                    src_ip_addr, dst_ip_addr, exp_port, verify_packet_any_port_lossy)
 
-        if isinstance(result, test.dataplane.PollSuccess):
-            if result.port != port:
-                # Flood case, check if packet rcvd on expected port as well
-                verify_packet(test, pkt, port)
-                return (-1, None)
-            else:
-                return (port, result.packet)
-
-        assert(isinstance(result, test.dataplane.PollFailure))
-        test.fail("Did not receive expected packet on any of ports %r for device %d.\n%s"
-                    % (ports, device_number, result.format()))
-        return (0, None)
+        return (matched_index, received)
 
 
     def send_rcv_ip_pkt_warm(self, in_port, sport, dport, src_ip_addr, dst_ip_addr,
-                       dst_port_list, ipv4=True, exp_port=None):
+                            exp_port, ipv4=True):
 
         # Simulate bidirectional traffic for mac learning, since mac learning(fdb) is flushed
         # as part of warm reboot
@@ -370,10 +437,10 @@ class FgEcmpTest(BaseTest):
 
         if ipv4:
             (matched_index, received) = self.send_rcv_ipv4_pkt(in_port, sport, dport,
-                    src_ip_addr, dst_ip_addr, dst_port_list, exp_port)
+                    src_ip_addr, dst_ip_addr, exp_port, verify_packet_warm)
         else:
             (matched_index, received) = self.send_rcv_ipv6_pkt(in_port, sport, dport,
-                    src_ip_addr, dst_ip_addr, dst_port_list, exp_port)
+                    src_ip_addr, dst_ip_addr, exp_port, verify_packet_warm)
 
         return (matched_index, received)
 
@@ -383,10 +450,10 @@ class FgEcmpTest(BaseTest):
 
         if ipv4:
             (matched_index, received) = self.send_rcv_ipv4_pkt(in_port, sport, dport,
-                    src_ip_addr, dst_ip_addr, dst_port_list)
+                    src_ip_addr, dst_ip_addr, dst_port_list, verify_packet_any_port)
         else:
             (matched_index, received) = self.send_rcv_ipv6_pkt(in_port, sport, dport,
-                    src_ip_addr, dst_ip_addr, dst_port_list)
+                    src_ip_addr, dst_ip_addr, dst_port_list, verify_packet_any_port)
 
         assert received
 
@@ -397,7 +464,7 @@ class FgEcmpTest(BaseTest):
 
 
     def send_rcv_ipv4_pkt(self, in_port, sport, dport,
-                         ip_src, ip_dst, dst_port_list, exp_port=None):
+                         ip_src, ip_dst, dst_port_list, verify_fn):
         src_mac = self.dataplane.get_mac(0, in_port)
         rand_int = random.randint(1, 254)
 
@@ -431,14 +498,11 @@ class FgEcmpTest(BaseTest):
         masked_exp_pkt.set_do_not_care_scapy(scapy.IP, "chksum")
         masked_exp_pkt.set_do_not_care_scapy(scapy.IP, "ttl")
 
-        if exp_port is None:
-            return verify_packet_any_port(self, masked_exp_pkt, dst_port_list)
-        else:
-            return self.verify_packet_warm(masked_exp_pkt, exp_port)
+        return verify_fn(self, masked_exp_pkt, dst_port_list)
 
 
     def send_rcv_ipv6_pkt(self, in_port, sport, dport,
-                         ip_src, ip_dst, dst_port_list, exp_port=None):
+                         ip_src, ip_dst, dst_port_list, verify_fn):
         src_mac = self.dataplane.get_mac(0, in_port)
         rand_int = random.randint(1, 254)
 
@@ -478,11 +542,7 @@ class FgEcmpTest(BaseTest):
         masked_exp_pkt.set_do_not_care_scapy(scapy.Ether, "src")
         masked_exp_pkt.set_do_not_care_scapy(scapy.IPv6, "hlim")
 
-        if exp_port is None:
-            return verify_packet_any_port(self, masked_exp_pkt, dst_port_list)
-        else:
-            return self.verify_packet_warm(masked_exp_pkt, exp_port)
-
+        return verify_fn(self, masked_exp_pkt,dst_port_list)
 
     def runTest(self):
         # Main function which triggers all the tests
